requirements_on_gadi.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.9" and python_version < "4.0"
aiohttp==3.12.13 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.9" and python_version < "3.11"
attrs==25.3.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2025.6.15 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
datasets==3.6.0 ; python_version >= "3.9" and python_version < "4.0"
dill==0.3.8 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.18.0 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.7.0 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2025.3.0 ; python_version >= "3.9" and python_version < "4.0"
fsspec[http]==2025.3.0 ; python_version >= "3.9" and python_version < "4.0"
hf-xet==1.1.5 ; python_version >= "3.9" and python_version < "4.0" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "arm64" or platform_machine == "aarch64")
huggingface-hub==0.33.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.9" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.6.2 ; python_version >= "3.9" and python_version < "4.0"
multiprocess==0.70.16 ; python_version >= "3.9" and python_version < "4.0"
networkx==3.2.1 ; python_version >= "3.9" and python_version < "4.0"
numpy==2.0.2 ; python_version >= "3.9" and python_version < "4.0"
nvidia-cublas-cu12==12.6.4.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cuda-cupti-cu12==12.6.80 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cuda-nvrtc-cu12==12.6.77 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cuda-runtime-cu12==12.6.77 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cudnn-cu12==9.5.1.17 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cufft-cu12==11.3.0.4 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cufile-cu12==1.11.1.6 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-curand-cu12==10.3.7.77 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cusolver-cu12==11.7.1.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cusparse-cu12==12.5.4.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-cusparselt-cu12==0.6.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-nccl-cu12==2.26.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-nvjitlink-cu12==12.6.85 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
nvidia-nvtx-cu12==12.6.77 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
pillow==11.2.1 ; python_version >= "3.9" and python_version < "4.0"
propcache==0.3.2 ; python_version >= "3.9" and python_version < "4.0"
pyarrow==20.0.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.9" and python_version < "4.0"
safetensors==0.5.3 ; python_version >= "3.9" and python_version < "4.0"
setuptools==80.9.0 ; python_version >= "3.12" and python_version < "4.0" or platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
sympy==1.14.0 ; python_version >= "3.9" and python_version < "4.0"
tokenizers==0.21.2 ; python_version >= "3.9" and python_version < "4.0"
torch==2.7.1 ; python_version >= "3.9" and python_version < "4.0"
torchvision==0.22.1 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.9" and python_version < "4.0"
transformers==4.53.0 ; python_version >= "3.9" and python_version < "4.0"
triton==3.3.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.9" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
xxhash==3.5.0 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.20.1 ; python_version >= "3.9" and python_version < "4.0"
